+++ docs/release-notes.md
@@ -31,6 +31,10 @@ open-source project, or update [test_pack.stbt_version] if you're using the
 
 ##### New features
 
+* `stbt.press` can be configured to use the ADB keypress mechanism from
+  `stbt.android.AdbDevice.press`. See the documentation for `--control` in the
+  [stbt(1) man page].
+
 ##### Minor fixes and packaging fixes